# docs/openapi/journal-service.json
{"openapi":"3.1.0","info":{"title":"Journal Service API","description":"Manages private journal entries, tags, full-text search, zip export, and Redis-cached stats.","version":"1.0"},"servers":[{"url":"http://localhost:8082","description":"Generated server url"}],"paths":{"/api/journal/entries/{id}":{"get":{"tags":["Journal Entries"],"summary":"Get entry by ID","operationId":"getEntry","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"400":{"description":"Bad Request","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/JournalEntryResponse"}}}}},"security":[{"bearerAuth":[]}]},"put":{"tags":["Journal Entries"],"summary":"Update entry","operationId":"updateEntry","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateEntryRequest"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/JournalEntryResponse"}}}}},"security":[{"bearerAuth":[]}]},"delete":{"tags":["Journal Entries"],"summary":"Soft-delete entry","operationId":"deleteEntry","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"400":{"description":"Bad Request","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"200":{"description":"OK"}},"security":[{"bearerAuth":[]}]}},"/api/journal/tags":{"get":{"tags":["Tags"],"summary":"List tags (alphabetical)","operationId":"listTags","responses":{"400":{"description":"Bad Request","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TagResponse"}}}}}},"security":[{"bearerAuth":[]}]},"post":{"tags":["Tags"],"summary":"Create tag","operationId":"createTag","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTagRequest"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TagResponse"}}}}},"security":[{"bearerAuth":[]}]}},"/api/journal/entries":{"get":{"tags":["Journal Entries"],"summary":"List entries (paginated)","operationId":"listEntries","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":31}},{"name":"from","in":"query","required":false,"schema":{"type":"string","format":"date"}},{"name":"to","in":"query","required":false,"schema":{"type":"string","format":"date"}},{"name":"q","in":"query","required":false,"schema":{"type":"string","maxLength":200,"minLength":0}}],"responses":{"400":{"description":"Bad Request","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageJournalEntryResponse"}}}}},"security":[{"bearerAuth":[]}]},"post":{"tags":["Journal Entries"],"summary":"Create entry","operationId":"createEntry","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateEntryRequest"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/JournalEntryResponse"}}}}},"security":[{"bearerAuth":[]}]}},"/api/journal/stats":{"get":{"tags":["Stats"],"summary":"Get journal stats for the current user","operationId":"getStats","responses":{"400":{"description":"Bad Request","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/StatsResponse"}}}}},"security":[{"bearerAuth":[]}]}},"/api/journal/entries/export":{"get":{"tags":["Journal Entries"],"summary":"Export entries as zip","operationId":"exportEntries","parameters":[{"name":"from","in":"query","required":true,"schema":{"type":"string","format":"date"}},{"name":"to","in":"query","required":true,"schema":{"type":"string","format":"date"}}],"responses":{"400":{"description":"Bad Request","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string","format":"byte"}}}}},"security":[{"bearerAuth":[]}]}},"/api/journal/entries/dates":{"get":{"tags":["Journal Entries"],"summary":"Get entry dates (lightweight — for calendar dot markers)","operationId":"getEntryDates","parameters":[{"name":"from","in":"query","required":true,"schema":{"type":"string","format":"date"}},{"name":"to","in":"query","required":true,"schema":{"type":"string","format":"date"}}],"responses":{"400":{"description":"Bad Request","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"type":"string"}}}}}},"security":[{"bearerAuth":[]}]}},"/api/journal/entries/date/{date}":{"get":{"tags":["Journal Entries"],"summary":"Get entry by specific date","operationId":"getEntryByDate","parameters":[{"name":"date","in":"path","required":true,"schema":{"type":"string","format":"date"}}],"responses":{"400":{"description":"Bad Request","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/JournalEntryResponse"}}}}},"security":[{"bearerAuth":[]}]}},"/api/journal/tags/{id}":{"delete":{"tags":["Tags"],"summary":"Delete tag","operationId":"deleteTag","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"400":{"description":"Bad Request","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"200":{"description":"OK"}},"security":[{"bearerAuth":[]}]}}},"components":{"schemas":{"ProblemDetail":{"type":"object","properties":{"type":{"type":"string","format":"uri"},"title":{"type":"string"},"status":{"type":"integer","format":"int32"},"detail":{"type":"string"},"instance":{"type":"string","format":"uri"},"properties":{"type":"object","additionalProperties":{}}}},"UpdateEntryRequest":{"type":"object","properties":{"title":{"type":"string","maxLength":255,"minLength":0},"content":{"type":"string","maxLength":10000,"minLength":0},"mood":{"type":"string","enum":["GREAT","GOOD","NEUTRAL","BAD","TERRIBLE"]},"tagIds":{"type":"array","items":{"type":"string","format":"uuid"},"maxItems":10,"minItems":0}},"required":["content"]},"JournalEntryResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"entryDate":{"type":"string","format":"date"},"title":{"type":"string"},"content":{"type":"string"},"mood":{"type":"string","enum":["GREAT","GOOD","NEUTRAL","BAD","TERRIBLE"]},"tags":{"type":"array","items":{"$ref":"#/components/schemas/TagResponse"}},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}}},"TagResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"name":{"type":"string"}}},"CreateTagRequest":{"type":"object","properties":{"name":{"type":"string","maxLength":50,"minLength":0}},"required":["name"]},"CreateEntryRequest":{"type":"object","properties":{"entryDate":{"type":"string","format":"date"},"title":{"type":"string","maxLength":255,"minLength":0},"content":{"type":"string","maxLength":10000,"minLength":0},"mood":{"type":"string","enum":["GREAT","GOOD","NEUTRAL","BAD","TERRIBLE"]},"tagIds":{"type":"array","items":{"type":"string","format":"uuid"},"maxItems":10,"minItems":0}},"required":["content","entryDate"]},"StatsResponse":{"type":"object","properties":{"totalEntries":{"type":"integer","format":"int64"},"totalWords":{"type":"integer","format":"int64"},"currentStreak":{"type":"integer","format":"int32"},"longestStreak":{"type":"integer","format":"int32"},"firstEntryDate":{"type":"string","format":"date"}}},"PageJournalEntryResponse":{"type":"object","properties":{"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/JournalEntryResponse"}},"number":{"type":"integer","format":"int32"},"first":{"type":"boolean"},"last":{"type":"boolean"},"numberOfElements":{"type":"integer","format":"int32"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"sort":{"$ref":"#/components/schemas/SortObject"},"empty":{"type":"boolean"}}},"PageableObject":{"type":"object","properties":{"offset":{"type":"integer","format":"int64"},"pageNumber":{"type":"integer","format":"int32"},"pageSize":{"type":"integer","format":"int32"},"sort":{"$ref":"#/components/schemas/SortObject"},"paged":{"type":"boolean"},"unpaged":{"type":"boolean"}}},"SortObject":{"type":"object","properties":{"empty":{"type":"boolean"},"sorted":{"type":"boolean"},"unsorted":{"type":"boolean"}}}},"securitySchemes":{"bearerAuth":{"type":"http","description":"RS256 JWT access token issued by the User Service","scheme":"bearer","bearerFormat":"JWT"}}}}