# services/journal-service/src/main/java/com/bravoscribe/journalservice/dto/CreateTagRequest.java
package com.bravoscribe.journalservice.dto;

import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.Size;

public record CreateTagRequest(
        @NotBlank @Size(max = 50) String name
) {}